Add files via upload
This command code is written for using the function code(predic5.m) for predicting the stream flow data. Using this command code the user can give data from multi number of stations as input and the code will give the predicted output result for all stations. This code can also give the results for different embedding dimensions from 2 to 10 and different number of neighbours from 1 to 10.